Phase 1 COMPLETE & Phase 2 In Progress: Fixed all 1812 CS compiler errors + 52 analyzer violations + Complete Guidebook with Build Enforcement (#349)
* Initial plan

* Phase 1 COMPLETE: Fixed 96 CS compiler errors (decimal/double type consistency)


* Phase 2: Fixed 16 analyzer violations (S109, CA1849, S6966, CA1031) + IFeatureBus compatibility


* Phase 1 MAJOR: Fixed 1646 CS0234 errors by renaming BotCore.Math → BotCore.Financial


* Phase 1: Fixed Side.FLAT enum errors (8 CS0117) - 1750/1812 CS errors fixed


* Phase 1 COMPLETE: Fixed final 62 CS errors - 0 compiler errors (100% complete)


* Phase 2 initiated: Analyzer violation assessment - 12,741 violations cataloged


* docs: Add hedge-fund-grade guardrails to Analyzer-Fix-Guidebook


* Phase 2 Round 85: Fix 12 Priority 1 violations (CA1031, S109) in risk management


* Phase 2 Round 86: Fix 24 Priority 1 violations (CA1031, S109) in trading feedback & ML config


* docs: Add solution-level build enforcement rails to Analyzer-Fix-Guidebook


---------

+++ src/BotCore/Integration/EpochFreezeEnforcement.cs
@@ -252,20 +252,20 @@ public sealed class EpochFreezeEnforcement
                 snapshot.ZoneAnchors["demand"] = new ZoneAnchor
                 {
                     ZoneType = "demand",
-                    AnchorPrice = snapshot.EntryPrice - features.distToDemandAtr * GetATRValue(snapshot.Symbol),
-                    DistanceATR = features.distToDemandAtr,
-                    BreakoutScore = features.breakoutScore,
-                    Pressure = features.zonePressure,
+                    AnchorPrice = snapshot.EntryPrice - (double)features.distToDemandAtr * GetATRValue(snapshot.Symbol),
+                    DistanceATR = (double)features.distToDemandAtr,
+                    BreakoutScore = (double)features.breakoutScore,
+                    Pressure = (double)features.zonePressure,
                     CapturedAt = DateTime.UtcNow
                 };
                 
                 snapshot.ZoneAnchors["supply"] = new ZoneAnchor
                 {
                     ZoneType = "supply",
-                    AnchorPrice = snapshot.EntryPrice + features.distToSupplyAtr * GetATRValue(snapshot.Symbol),
-                    DistanceATR = features.distToSupplyAtr,
-                    BreakoutScore = features.breakoutScore,
-                    Pressure = features.zonePressure,
+                    AnchorPrice = snapshot.EntryPrice + (double)features.distToSupplyAtr * GetATRValue(snapshot.Symbol),
+                    DistanceATR = (double)features.distToSupplyAtr,
+                    BreakoutScore = (double)features.breakoutScore,
+                    Pressure = (double)features.zonePressure,
                     CapturedAt = DateTime.UtcNow
                 };
             }
@@ -319,7 +319,7 @@ public sealed class EpochFreezeEnforcement
             // Validate demand zone anchor
             if (snapshot.ZoneAnchors.TryGetValue("demand", out var demandAnchor))
             {
-                var currentDemandPrice = request.CurrentPrice - currentFeatures.distToDemandAtr * currentATR;
+                var currentDemandPrice = request.CurrentPrice - (double)currentFeatures.distToDemandAtr * currentATR;
                 var priceDifference = Math.Abs(currentDemandPrice - demandAnchor.AnchorPrice);
                 var toleranceTicks = _config.PriceToleranceTicks * GetTickSize(snapshot.Symbol);
                 
@@ -339,7 +339,7 @@ public sealed class EpochFreezeEnforcement
             // Validate supply zone anchor
             if (snapshot.ZoneAnchors.TryGetValue("supply", out var supplyAnchor))
             {
-                var currentSupplyPrice = request.CurrentPrice + currentFeatures.distToSupplyAtr * currentATR;
+                var currentSupplyPrice = request.CurrentPrice + (double)currentFeatures.distToSupplyAtr * currentATR;
                 var priceDifference = Math.Abs(currentSupplyPrice - supplyAnchor.AnchorPrice);
                 var toleranceTicks = _config.PriceToleranceTicks * GetTickSize(snapshot.Symbol);
